Revert "Build JNI IFA switching helpers as C calls"
