+++ runtime/codert_vm/armnathelp.m4
@@ -366,6 +366,8 @@ UNUSED(jitNewPackedArray)
 UNUSED(jitResolvePackedArrayFieldLength)
 UNUSED(jitResolveIsPackedFieldNested)
 UNUSED(jitNewObjectNoTenantInit)
+UNUSED(jitPostJNICallOffloadCheck)
+UNUSED(jitPreJNICallOffloadCheck)
 UNUSED(jitFindFieldSignatureClass)
 UNUSED(icallVMprJavaSendInvokeWithArgumentsHelperL)
 UNUSED(j2iInvokeWithArguments)
