+++ runtime/codert_vm/pnathelp.m4
@@ -411,6 +411,8 @@ UNUSED(jitNewPackedArray)
 UNUSED(jitResolvePackedArrayFieldLength)
 UNUSED(jitResolveIsPackedFieldNested)
 UNUSED(jitNewObjectNoTenantInit)
+UNUSED(jitPostJNICallOffloadCheck)
+UNUSED(jitPreJNICallOffloadCheck)
 UNUSED(jitFindFieldSignatureClass)
 UNUSED(icallVMprJavaSendInvokeWithArgumentsHelperL)
 UNUSED(j2iInvokeWithArguments)
