+++ runtime/codert_vm/xnathelp.m4
@@ -465,6 +465,8 @@ UNUSED(jitNewPackedArray)
 UNUSED(jitResolvePackedArrayFieldLength)
 UNUSED(jitResolveIsPackedFieldNested)
 UNUSED(jitNewObjectNoTenantInit)
+UNUSED(jitPostJNICallOffloadCheck)
+UNUSED(jitPreJNICallOffloadCheck)
 UNUSED(jitFindFieldSignatureClass)
 UNUSED(icallVMprJavaSendInvokeWithArgumentsHelperL)
 UNUSED(j2iInvokeWithArguments)
