+++ runtime/compiler/z/codegen/J9CodeGenerator.cpp
@@ -39,13 +39,13 @@
 #include "env/jittypes.h"
 #include "il/Node.hpp"
 #include "il/Node_inlines.hpp"
-#include "z/codegen/J9S390JNILinkage.hpp"
 #include "z/codegen/J9S390PrivateLinkage.hpp"
 #include "z/codegen/J9S390SystemLinkage.hpp"
 #include "z/codegen/J9S390CHelperLinkage.hpp"
 #include "z/codegen/S390GenerateInstructions.hpp"
 #include "z/codegen/S390Recompilation.hpp"
 #include "z/codegen/S390Register.hpp"
+#include "z/codegen/J9S390PrivateLinkage.hpp"
 #include "z/codegen/ReduceSynchronizedFieldLoad.hpp"
 
 #define OPT_DETAILS "O^O CODE GENERATION: "
