+++ runtime/compiler/z/codegen/J9S390SystemLinkage.hpp
@@ -98,10 +98,6 @@ public:
    virtual int32_t storeExtraEnvRegForBuildArgs(TR::Node * callNode, TR::Linkage* linkage, TR::RegisterDependencyConditions * dependencies,
          bool isFastJNI, int32_t stackOffset, int8_t gprSize, uint32_t &numIntegerArgs);
 
-   /**
-    * \brief
-    * Insert NOP padding and call descriptor pseudo instruction after a XPLINK call.
-    */
    TR::Instruction * genCallNOPAndDescriptor(TR::Instruction * cursor, TR::Node *node, TR::Node *callNode, TR_XPLinkCallTypes callType);
    };
